2021-01-17 - Housekeeping
Created Base Class "BaseDSClientSessionCleanup". Enter-DSClientSession and Exit-DSClientSession both derive from this class, thus reducing code duplication for cleaning up session data.
