+++ PSAsigraDSClient/EnterDSClientSession.cs
@@ -6,11 +6,11 @@ using static PSAsigraDSClient.DSClientCommon;
 namespace PSAsigraDSClient
 {
     [Cmdlet(VerbsCommon.Enter, "DSClientSession")]
-    public class EnterDSClientSession: PSCmdlet
+    public class EnterDSClientSession: BaseDSClientSessionCleanup
     {
-        [Parameter(Position = 0, HelpMessage = "Specify the DSClient Host to connect to")]
+        [Parameter(Position = 0, HelpMessage = "Specify the DS-Client Host to connect to")]
         [ValidateNotNullOrEmpty]
-        public string Host { get; set; }
+        public new string Host { get; set; }
 
         [Parameter(Position = 1, HelpMessage = "Specify the TCP port to connect to")]
         [ValidateRange(1, 65535)]
@@ -19,106 +19,17 @@ namespace PSAsigraDSClient
         [Parameter(Position = 3, HelpMessage = "Specify to NOT establish an SSL Connection")]
         public SwitchParameter NoSSL { get; set; }
 
-        [Parameter(Position = 4, HelpMessage = "Specify the Asigra DSClient API Version to use")]
+        [Parameter(Position = 4, HelpMessage = "Specify the Asigra DS-Client API Version to use")]
         public string APIVersion { get; set; } = "13.0.0.0";
 
         [Parameter(Position = 5, HelpMessage = "Specify Credentials to use to connect to DSClient")]
         [ValidateNotNullOrEmpty]
         public PSCredential Credential { get; set; }
 
-        protected override void ProcessRecord()
+        protected override void ProcessCleanSession()
         {
             string user = Credential.UserName;
             string pwd = Credential.GetNetworkCredential().Password;
-
-            // Remove any previous DataType from SessionState and store the current Backup Set DataType
-            SessionState.PSVariable.Remove("RestoreType");
-
-            // Check for a previous Backup Set Restore View stored in Session State
-            WriteVerbose("Performing Action: Check for previous DS-Client Validation View Sessions");
-            BackupSetRestoreView previousRestoreSession = SessionState.PSVariable.GetValue("RestoreView", null) as BackupSetRestoreView;
-
-            // If a previous session is found, remove it
-            if (previousRestoreSession != null)
-            {
-                WriteVerbose("Notice: Previous Restore View found");
-                try
-                {
-                    WriteVerbose("Performing Action: Dispose Restore View");
-                    previousRestoreSession.Dispose();
-                }
-                catch
-                {
-                    WriteVerbose("Notice: Previous Session failed to Dispose");
-                }
-                WriteVerbose("Performing Action: Remove on Restore View Session");
-                SessionState.PSVariable.Remove("RestoreView");
-            }
-
-            // Check for a previous Backup Set Validation View stored in Session State
-            WriteVerbose("Performing Action: Check for previous DS-Client Delete View Sessions");
-            BackupSetDeleteView previousDeleteSession = SessionState.PSVariable.GetValue("DeleteView", null) as BackupSetDeleteView;
-
-            // If a previous session is found, remove it
-            if (previousDeleteSession != null)
-            {
-                WriteVerbose("Notice: Previous Delete View found");
-                try
-                {
-                    WriteVerbose("Performing Action: Dispose Delete View");
-                    previousDeleteSession.Dispose();
-                }
-                catch
-                {
-                    WriteVerbose("Notice: Previous Session failed to Dispose");
-                }
-                WriteVerbose("Performing Action: Remove on Delete View Session");
-                SessionState.PSVariable.Remove("DeleteView");
-            }
-
-            // Check for a previous Backup Set Validation View stored in Session State
-            WriteVerbose("Performing Action: Check for previous DS-Client Validation View Sessions");
-            BackupSetValidationView previousValidationSession = SessionState.PSVariable.GetValue("ValidateView", null) as BackupSetValidationView;
-
-            // If a previous session is found, remove it
-            if (previousValidationSession != null)
-            {
-                WriteVerbose("Notice: Previous Validation View found");
-                try
-                {
-                    WriteVerbose("Performing Action: Dispose Validation View");
-                    previousValidationSession.Dispose();
-                }
-                catch
-                {
-                    WriteVerbose("Notice: Previous Session failed to Dispose");
-                }
-                WriteVerbose("Performing Action: Remove on Valation View Session");
-                SessionState.PSVariable.Remove("ValidateView");
-            }
-
-            WriteVerbose("Performing Action: Check for existing DS-Client Sessions");
-            ClientConnection previousSession = SessionState.PSVariable.GetValue("DSClientSession", null) as ClientConnection;
-
-            if (previousSession != null)
-            {
-                WriteVerbose("Notice: Previous DS-Client Session found");
-                try
-                {
-                    WriteVerbose("Performing Action: Logout on DS-Client Session");
-                    previousSession.logout();
-                    WriteVerbose("Performing Action: Dispose DS-Client Session");
-                    previousSession.Dispose();
-                }
-                catch
-                {
-                    WriteVerbose("Notice: Previous session failed to dispose");                    
-                }
-                WriteVerbose("Performing Action: Remove on DS-Client Session");
-                SessionState.PSVariable.Remove("DSClientSession");
-                SessionState.PSVariable.Remove("DSClientOSType");
-                WriteObject("DS-Client Session removed.");
-            }
 
             WriteVerbose("Performing Action: Establish DS-Client Session");
             ClientConnection DSClientSession = ConnectSession(Host, Port, NoSSL, APIVersion, user, pwd);
